+++ data/commits.json
@@ -1,16 +1,1 @@
-{
-  "Foobar": {
-    "foobar": [
-      {
-        "message": "Update README.md",
-        "author": {
-          "name": "Foobar",
-          "email": "foobar@email.com",
-          "date": "2017-02-16T19:25:15Z"
-        },
-        "url": "https://github.com/Foobar/foobar/commit/20ad02f81024c7dc9bff6da3d8e996725203923e",
-        "sha": "20ad02f81024c7dc9bff6da3d8e996725203923e"
-      }
-    ]
-  }
-}
+[{"sha":"61d13b343c0e75de78be0ee55ed7250987563ee2","commit_message":"webhooks","author":{"name":"Egle Libby","email":"eglelibby@gmail.com","date":"2017-03-30T21:37:30Z"},"html_url":"https://github.com/eglital/project_github_commit_feed/commit/61d13b343c0e75de78be0ee55ed7250987563ee2"},{"sha":"3321bd6efd58f2a9d71b7e16306bf5097d9ee0cc","commit_message":"update webhooks response","author":{"name":"Egle Libby","email":"eglelibby@gmail.com","date":"2017-03-30T20:47:11Z"},"html_url":"https://github.com/eglital/project_github_commit_feed/commit/3321bd6efd58f2a9d71b7e16306bf5097d9ee0cc"},{"sha":"0c486f4081d73b787b79794a3888b2cd6e7df6fb","commit_message":"enabled webhooks","author":{"name":"William Whitworth","email":"william10charles@gmail.com","date":"2017-03-30T20:41:10Z"},"html_url":"https://github.com/eglital/project_github_commit_feed/commit/0c486f4081d73b787b79794a3888b2cd6e7df6fb"},{"sha":"a0e7c88263630237b815cfe253cd154a0f4d222c","commit_message":"Please work","author":{"name":"William Whitworth","email":"william10charles@gmail.com","date":"2017-03-30T20:23:37Z"},"html_url":"https://github.com/eglital/project_github_commit_feed/commit/a0e7c88263630237b815cfe253cd154a0f4d222c"},{"sha":"e680c379e25e59eba5bd1f95399818d217ba3010","commit_message":"adfjadlfk","author":{"name":"William Whitworth","email":"william10charles@gmail.com","date":"2017-03-30T20:20:55Z"},"html_url":"https://github.com/eglital/project_github_commit_feed/commit/e680c379e25e59eba5bd1f95399818d217ba3010"},{"sha":"89d454f84ba1d32cd820c89c3a7de87c11a4aedd","commit_message":"new test again","author":{"name":"William Whitworth","email":"william10charles@gmail.com","date":"2017-03-30T20:17:52Z"},"html_url":"https://github.com/eglital/project_github_commit_feed/commit/89d454f84ba1d32cd820c89c3a7de87c11a4aedd"},{"sha":"78a4cdfe1946f52c1d15febf54ca7c73803e3277","commit_message":"new test","author":{"name":"William Whitworth","email":"william10charles@gmail.com","date":"2017-03-30T20:13:44Z"},"html_url":"https://github.com/eglital/project_github_commit_feed/commit/78a4cdfe1946f52c1d15febf54ca7c73803e3277"},{"sha":"2ab462c706667837c1ec50ed0c763bf5d2276d24","commit_message":"test","author":{"name":"William Whitworth","email":"william10charles@gmail.com","date":"2017-03-30T20:12:03Z"},"html_url":"https://github.com/eglital/project_github_commit_feed/commit/2ab462c706667837c1ec50ed0c763bf5d2276d24"},{"sha":"20a480647e16ffdccaeeff6abbb7a26b0246cd06","commit_message":"deleted files?","author":{"name":"William Whitworth","email":"william10charles@gmail.com","date":"2017-03-30T20:05:02Z"},"html_url":"https://github.com/eglital/project_github_commit_feed/commit/20a480647e16ffdccaeeff6abbb7a26b0246cd06"},{"sha":"ab2557bc431cc258db7c4884490ce934634bdf0a","commit_message":"update gitignore","author":{"name":"Egle Libby","email":"eglelibby@gmail.com","date":"2017-03-30T18:06:34Z"},"html_url":"https://github.com/eglital/project_github_commit_feed/commit/ab2557bc431cc258db7c4884490ce934634bdf0a"},{"sha":"7218c876db76c61e0c30bb43da73ca9777c7b3fc","commit_message":"scrubbed data","author":{"name":"Egle Libby","email":"eglelibby@gmail.com","date":"2017-03-30T18:01:26Z"},"html_url":"https://github.com/eglital/project_github_commit_feed/commit/7218c876db76c61e0c30bb43da73ca9777c7b3fc"},{"sha":"2938a0d34f3df38b88bbd9ddc407603cf02daf6c","commit_message":"forgot to save","author":{"name":"William Whitworth","email":"william10charles@gmail.com","date":"2017-03-30T17:33:00Z"},"html_url":"https://github.com/eglital/project_github_commit_feed/commit/2938a0d34f3df38b88bbd9ddc407603cf02daf6c"},{"sha":"11b3fd27ac683310f34e8aeb482f3035b064ea6c","commit_message":"used wrapper to dispaly commits","author":{"name":"William Whitworth","email":"william10charles@gmail.com","date":"2017-03-30T17:31:44Z"},"html_url":"https://github.com/eglital/project_github_commit_feed/commit/11b3fd27ac683310f34e8aeb482f3035b064ea6c"},{"sha":"ae822d7145906ae7a46b7aa637c8dc5ad79dbccf","commit_message":"serving commit","author":{"name":"Egle Libby","email":"eglelibby@gmail.com","date":"2017-03-30T17:03:07Z"},"html_url":"https://github.com/eglital/project_github_commit_feed/commit/ae822d7145906ae7a46b7aa637c8dc5ad79dbccf"},{"sha":"0c4954e72095953db4dd668537b93cef0dd50624","commit_message":"created router?","author":{"name":"William Whitworth","email":"william10charles@gmail.com","date":"2017-03-30T16:32:30Z"},"html_url":"https://github.com/eglital/project_github_commit_feed/commit/0c4954e72095953db4dd668537b93cef0dd50624"},{"sha":"19e64459506ba682797990f4399675804d73ca9c","commit_message":"replaced the commtFeed with JSON data","author":{"name":"William Whitworth","email":"william10charles@gmail.com","date":"2017-03-30T16:10:45Z"},"html_url":"https://github.com/eglital/project_github_commit_feed/commit/19e64459506ba682797990f4399675804d73ca9c"},{"sha":"18126038b75790730911a20c00caa1768165023c","commit_message":"server setup","author":{"name":"Egle Libby","email":"eglelibby@gmail.com","date":"2017-03-30T16:02:22Z"},"html_url":"https://github.com/eglital/project_github_commit_feed/commit/18126038b75790730911a20c00caa1768165023c"},{"sha":"fdb9dd7d7fd067da4cc6acd4edfc38080f9b7bb5","commit_message":"user interface","author":{"name":"William Whitworth","email":"william10charles@gmail.com","date":"2017-03-30T15:32:31Z"},"html_url":"https://github.com/eglital/project_github_commit_feed/commit/fdb9dd7d7fd067da4cc6acd4edfc38080f9b7bb5"},{"sha":"a3ce231d8676dd22d77b5083005ae79938ee89a5","commit_message":"finish github wrapper","author":{"name":"Egle Libby","email":"eglelibby@gmail.com","date":"2017-03-30T15:01:43Z"},"html_url":"https://github.com/eglital/project_github_commit_feed/commit/a3ce231d8676dd22d77b5083005ae79938ee89a5"},{"sha":"2a34d0b6a23e171d8644bbb361161c5a827492ff","commit_message":"initial set up","author":{"name":"William Whitworth","email":"william10charles@gmail.com","date":"2017-03-30T14:34:00Z"},"html_url":"https://github.com/eglital/project_github_commit_feed/commit/2a34d0b6a23e171d8644bbb361161c5a827492ff"},{"sha":"ceec1b2306888905ae12750558706174132a6741","commit_message":"Update README.md","author":{"name":"Chris Scavello","email":"bideowego@gmail.com","date":"2017-02-21T16:48:57Z"},"html_url":"https://github.com/eglital/project_github_commit_feed/commit/ceec1b2306888905ae12750558706174132a6741"},{"sha":"526145a8b50b13bafc85f9c8e3fa603c93912f55","commit_message":"Initial commit","author":{"name":"Chris Scavello","email":"bideowego@gmail.com","date":"2017-02-21T16:47:14Z"},"html_url":"https://github.com/eglital/project_github_commit_feed/commit/526145a8b50b13bafc85f9c8e3fa603c93912f55"}]
